# docs/swagger.yaml
basePath: /
definitions:
  entity.Channel:
    enum:
    - telegram
    - email
    type: string
    x-enum-varnames:
    - Telegram
    - Email
  entity.Notification:
    properties:
      channel:
        $ref: '#/definitions/entity.Channel'
      createdAt:
        type: string
      id:
        type: string
      lastError:
        type: string
      payload:
        type: string
      retryCount:
        type: integer
      scheduledAt:
        type: string
      sentAt:
        type: string
      status:
        $ref: '#/definitions/entity.Status'
      userID:
        type: string
    type: object
  entity.Status:
    enum:
    - waiting
    - in_process
    - sent
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusWaiting
    - StatusInProcess
    - StatusSent
    - StatusFailed
    - StatusCancelled
  handler.CreateNotificationRequest:
    properties:
      channel:
        allOf:
        - $ref: '#/definitions/entity.Channel'
        enum:
        - telegram
        - email
      payload:
        maxLength: 100000
        type: string
      scheduled_at:
        type: string
      user_id:
        type: string
    required:
    - channel
    - payload
    - scheduled_at
    - user_id
    type: object
  handler.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      details:
        example: notify with id 123 does not exist
        type: string
      error:
        example: notify not found
        type: string
    type: object
  handler.LinkTokenResponse:
    properties:
      expires_in:
        type: string
      link:
        type: string
      message:
        type: string
      token:
        type: string
    required:
    - expires_in
    - link
    - token
    type: object
  handler.NotificationCreatedResponse:
    properties:
      id:
        type: string
      message:
        type: string
    required:
    - id
    type: object
  handler.RegisterUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - email
    - name
    type: object
  handler.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
  handler.UserRegisteredResponse:
    properties:
      message:
        type: string
      user_id:
        type: string
    required:
    - user_id
    type: object
host: localhost:8080
info:
  contact:
    email: stormkillpeople@gmail.com
    name: RidusM
  description: API for working with notifications
  license:
    name: MIT-0
    url: https://github.com/aws/mit-0
  termsOfService: http://swagger.io/terms/
  title: Notification Service API
  version: "1.0"
paths:
  /health:
    get:
      description: Return service status and current timestamp. No authentication
        required.
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - System
  /notify:
    post:
      consumes:
      - application/json
      description: Schedules a notification to be sent to a specific user at a given
        time
      parameters:
      - description: Notification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Notification created
          schema:
            $ref: '#/definitions/handler.NotificationCreatedResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a scheduled notification
      tags:
      - Notifications
  /notify/{id}:
    delete:
      consumes:
      - application/json
      description: Cancels a scheduled notification if it hasn't been sent yet
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cancellation successful
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Notification already sent or cancelled
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Cancel a notification
      tags:
      - Notifications
    get:
      consumes:
      - application/json
      description: Returns the current status of a notification by its ID
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification details
          schema:
            $ref: '#/definitions/entity.Notification'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get notification status
      tags:
      - Notifications
  /users:
    post:
      consumes:
      - application/json
      description: Registers a user to receive notifications via Email or Telegram
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/handler.UserRegisteredResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - Users
  /users/{user_id}/link-token:
    post:
      consumes:
      - application/json
      description: Generates a one-time token to link the user's account with Telegram
        bot
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Link token and instruction
          schema:
            $ref: '#/definitions/handler.LinkTokenResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Generate Telegram Link Token
      tags:
      - Users
swagger: "2.0"